Convert SSB and Rands Winrate chat plugins to tsx (#11030)
