+++ server/chat-jsx.tsx
@@ -29,6 +29,7 @@ export interface PSElements extends preact.JSX.IntrinsicElements {
 	username: { name?: string, class?: string, children?: preact.VNode | string };
 	psicon: { pokemon: string } | { item: string } | { type: string } | { category: string };
 	center: { class?: string };
+	font: { size?: string, color?: string };
 }
 
 export { render };
